menambahkan fitur2 kecil

+++ resources/views/admin/content/users.blade.php
@@ -35,16 +35,22 @@
                     <tbody>
                       @foreach($users as $usr)
                       <tr>
-                        <td>{{ $usr->namaUser }}</td>
+                        <td>{{ $usr->namaUser }} 
+@if($usr->idBidang == 15)
+ <sup class="red-text text-darken-2" >Admin</sup>
+@endif
+</td>
                         <td>{{ $usr->emailUser }}</td>
                         <td>{{ $usr->namaBidang }}</td>
                         <td>
                           <a href="/admin/edit-data-pengguna/{{ $usr->idUser }}" class="mb-6 btn-floating waves-effect waves-light purple lightrn-1">
                             <i class="material-icons">edit</i>
                           </a>
+@if($usr->idBidang != 15)
                           <a href="/admin/delete-data-pengguna/{{ $usr->idUser }}" id="tombol-delete" class="mb-6 btn-floating waves-effect waves-light purple lightrn-1 ">
                             <i class="material-icons">delete_forever</i>
                           </a>
+@endif
                         </td>
                       </tr>
                       @endforeach
